feat: add routing and 404 page
