+++ src/App.tsx
@@ -1,29 +1,16 @@
-import { useState } from 'react';
-import Search from './Search';
-import PokemonList from './PokemonList';
+import React from 'react';
+import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
+import Home from './components/Home';
+import NotFound from './components/NotFound';
 
 const App: React.FC = () => {
-  const [searchTerm, setSearchTerm] = useState<string>('');
-
-  const handleSearch = (term: string) => {
-    setSearchTerm(term);
-  };
-
-  const throwError = () => {
-    throw new Error('Test error');
-  };
-
   return (
-    <div className="App">
-      <h1>My Pokemon App</h1>
-      <div className="top-section">
-        <Search onSearch={handleSearch} />
-      </div>
-      <div className="bottom-section">
-        <PokemonList searchTerm={searchTerm} />
-      </div>
-      <button onClick={throwError}>Throw Error</button>
-    </div>
+    <Router>
+      <Routes>
+        <Route path="/main" element={<Home />} />
+        <Route path="*" element={<NotFound />} />
+      </Routes>
+    </Router>
   );
 };
