+++ src/components/NotFound.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const NotFound: React.FC = () => {
+  return (
+    <div>
+      <h1>404 - Page Not Found</h1>
+      <p>The page you are looking for does not exist.</p>
+    </div>
+  );
+};
+
+export default NotFound;
